src/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
lxml==5.3.1
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
